polls_admin/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
supabase>=2.0.0
python-dotenv==1.0.1
Jinja2==3.1.4
itsdangerous>=2.0
python-multipart>=0.0.6
